+++ spec/swagger_v2/api_swagger_v2_mounted_spec.rb
@@ -81,6 +81,7 @@ describe 'swagger spec v2.0' do
           requires :id, type: Integer
         end
         delete '/dummy/:id' do
+          {}
         end
 
         namespace :other_thing do