+++ spec/swagger_v2/float_api_spec.rb
@@ -11,6 +11,7 @@ describe 'Float Params' do
         requires :a_float, type: Float
       end
       post :splines do
+        { message: 'hi' }
       end
 
       add_swagger_documentation